Short term shrine reorg
